src/main/persistence/JsonReader.java
package persistence;

import model.Counter;
import org.json.JSONObject;

import java.io.IOException;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.stream.Stream;

public class JsonReader {
    private String source;
    private Counter counter = new Counter();

    public JsonReader(String source) {
        this.source = source;
    }

    public int readCounter() throws IOException {
        String jsonData = readFile(source);
        JSONObject jsonObject = new JSONObject(jsonData);
        return parseCounter(jsonObject);
    }

    private int parseCounter(JSONObject jsonObject) {
        int count = jsonObject.getInt("count");
        return count;
    }

    private String readFile(String source) throws IOException {
        StringBuilder contentBuilder = new StringBuilder();

        try (Stream<String> stream = Files.lines(Paths.get(source), StandardCharsets.UTF_8)) {
            stream.forEach(s -> contentBuilder.append(s));
        }

        return contentBuilder.toString();
    }
}
